server/requirements.txt
fastapi==0.95.1
python-socketio==5.8.0
uvicorn==0.21.1
